requirements.txt
beautifulsoup4==4.14.2
bleach==6.2.0
bs4==0.0.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
dotenv==0.9.9
duckdb==1.4.1
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-core==2.5.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
idna==3.11
kaggle==1.7.4.5
nba_api==1.10.2
numpy==2.3.4
pandas==2.3.3
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
wget==3.2
